Commands/requirements.txt
mysql-connector-python==8.1.0
pandas==1.5.2
python-dotenv==1.0.0
Requests==2.31.0
